requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.120.0
python-dotenv==1.0.1
